+++ tests/test_sequence.py
@@ -3,7 +3,7 @@ import unittest
 
 from pkg_resources import Requirement, resource_filename
 
-from emoji_data import (EmojiSequence, QualifiedType, code_points_to_string,
+from emoji_data import (EmojiCharacter, EmojiSequence, QualifiedType, code_points_to_string,
                         detect_qualified, is_emoji_character,
                         is_emoji_flag_sequence, is_emoji_keycap_sequence,
                         is_emoji_modifier_sequence)
@@ -23,6 +23,14 @@ class SequenceTestCase(unittest.TestCase):
         with open(DATAFILE_TEST, encoding='utf-8') as fp:
             for content, _ in read_data_file_iterable(fp):
                 cls.test_data.append(tuple(s.strip() for s in content.split(';', 1)))  # type: ignore
+
+    def setUp(self):
+        EmojiCharacter.initial()
+        EmojiSequence.initial()
+
+    def tearDown(self):
+        EmojiSequence.release()
+        EmojiCharacter.release()
 
     def test_len(self):
         for code_points, *_ in self.test_data:
@@ -60,11 +68,14 @@ class SequencePatternTestCase(unittest.TestCase):
                 'abc',
                 ' abc\n bcd'
         ):
-            self.assertIsNone(EmojiSequence.pattern.match(s))
+            self.assertListEqual(EmojiSequence.find(s), [])
 
     def test_single_one_char_emoji(self):
         s = '😀'
-        self.assertIsNotNone(EmojiSequence.pattern.match(s))
+        self.assertListEqual(
+            [m.string for m, _, _ in EmojiSequence.find(s)],
+            [s]
+        )
 
     def test_one_char_emoji_started_text(self):
         s = '😀笑脸'
@@ -74,17 +85,13 @@ class SequencePatternTestCase(unittest.TestCase):
     def test_1char_emoji_in_text(self):
         s = '这是😀笑脸'
         self.assertIsNone(EmojiSequence.pattern.match(s))
-        m = EmojiSequence.pattern.search(s)
-        self.assertIsNotNone(m)
-        self.assertEqual(m.span(), (2, 3))
+        e, p0, p1 = EmojiSequence.find(s)[0]
+        self.assertEqual((p0, p1), (2, 3))
 
     def test_many_one_char_emoji_in_text(self):
         s = '1😛 2😛 3😛'
         cnt = 0
-        m = EmojiSequence.pattern.search(s)
-        while m:
-            cnt += 1
-            m = EmojiSequence.pattern.search(s, m.end())
+        cnt = sum(1 for _ in EmojiSequence.iter_find(s))
         self.assertEqual(cnt, 3)
 
     def test_single_multi_chars_emoji(self):
@@ -99,26 +106,17 @@ class SequencePatternTestCase(unittest.TestCase):
     def test_single_multi_chars_emoji_in_text(self):
         s = '这个☺️也是笑脸'
         self.assertIsNone(EmojiSequence.pattern.match(s))
-        m = EmojiSequence.pattern.search(s)
-        self.assertIsNotNone(m)
-        self.assertEqual(m.span(), (2, 4))
+        e, p0, p1 = EmojiSequence.find(s)[0]
+        self.assertEqual((p0, p1), (2, 4))
 
     def test_many_multi_chars_emoji_in_text(self):
         s = '1☺️ 2☺️ 3☺️'
-        cnt = 0
-        m = EmojiSequence.pattern.search(s)
-        while m:
-            cnt += 1
-            m = EmojiSequence.pattern.search(s, m.end())
+        cnt = sum(1 for _ in EmojiSequence.iter_find(s))
         self.assertEqual(cnt, 3)
 
     def test_many_multi_chars_emoji_in_multiline_text(self):
         s = '1☺️ {0}2☺️ {0}3☺️'.format(os.linesep)
-        cnt = 0
-        m = EmojiSequence.pattern.search(s)
-        while m:
-            cnt += 1
-            m = EmojiSequence.pattern.search(s, m.end())
+        cnt = sum(1 for _ in EmojiSequence.iter_find(s))
         self.assertEqual(cnt, 3)
 
     def test_continues_two_with_same_start_part(self):
@@ -127,11 +125,7 @@ class SequencePatternTestCase(unittest.TestCase):
             '1F468 200D 1F468 200D 1F467 200D 1F467'
         ]
         s = ''.join(code_points_to_string(m) for m in emojis)
-        cnt = 0
-        m = EmojiSequence.pattern.search(s)
-        while m:
-            cnt += 1
-            m = EmojiSequence.pattern.search(s, m.end())
+        cnt = sum(1 for _ in EmojiSequence.iter_find(s))
         self.assertEqual(cnt, 2)